clipboard: Improve interactions, colors, and loading states
   - Text items paste on click, media items show menu on click
   - Rename pin/unpin to add/remove favorites for consistency
   - Use themed accent colors for selected category and token items
   - Add loading state for tokenized clipboard (fixes first-use "no tokens" issue)
   - Fix 3 Kotlin compiler warnings
   - Add translations for loading state to all languages

+++ app/src/main/java/org/fcitx/fcitx5/android/ui/main/settings/behavior/SplitKeyboardCalibrationActivity.kt
@@ -449,7 +449,7 @@ class SplitKeyboardCalibrationActivity : AppCompatActivity() {
                     keyboardPrefs.keyboardHeightPercentLandscape.getValue()
                 else -> keyboardPrefs.keyboardHeightPercent.getValue()
             }
-            val keyboardHeight = (screenHeight * heightPercent / 100).toInt()
+            val keyboardHeight = screenHeight * heightPercent / 100
 
             val sidePaddingDp = when (resources.configuration.orientation) {
                 Configuration.ORIENTATION_LANDSCAPE ->
